Refactor package.json creation
