+++ src/package/index.js
@@ -1,4 +1,5 @@
 import Generator from 'yeoman-generator'
+import {sortObj} from '../utils'
 
 module.exports = class Package extends Generator {
   prompting () {
@@ -22,7 +23,17 @@ module.exports = class Package extends Generator {
 
   writing () {
     /* Set basic info */
-    let info = {
+    let info = this.getBaseInfo()
+
+    let tpl = this.fs.readJSON(this.templatePath('package.json'))
+    info = Object.assign(info, tpl)
+    info = projectAdjustments(info, this.props, this.config.get('projectType'))
+
+    this.fs.writeJSON(this.destinationPath('package.json'), info)
+  }
+
+  getBaseInfo () {
+    return {
       name: this.config.get('name'),
       description: this.config.get('description'),
       version: this.config.get('version'),
@@ -31,27 +42,10 @@ module.exports = class Package extends Generator {
         email: this.user.git.email()
       }
     }
-
-    let tpl = this.fs.readJSON(this.templatePath('package.json'))
-    info = Object.assign(info, tpl)
-    info = projectAdjustments(info, this.props, this.config.get('projectType'))
-
-    this.fs.writeJSON(this.destinationPath('package.json'), info)
   }
 }
 
 /* Helper functions */
-function sortObj (obj) {
-  let keys = Object.keys(obj).sort()
-  let sorted = {}
-
-  keys.map((key) => {
-    sorted = Object.assign(sorted, {[key]: obj[key]})
-  })
-
-  return sorted
-}
-
 function projectAdjustments (info, props, type) {
   switch (type) {
     case 'static-site':
@@ -64,62 +58,38 @@ function projectAdjustments (info, props, type) {
 }
 
 function typeStaticSite (info, props) {
-  let result = Object.assign({}, info)
-
-  let scripts = Object.assign(result.scripts, {
-    'build:prod': 'gulp build:prod',
-    'test:ui': 'babel-node test/ui'
+  let scripts = Object.assign(info.scripts, {
+    'build:prod': 'gulp build:prod'
   })
 
-  let globals = [
-    ...result.standard.globals,
-    'fixture',
-    'test'
-  ].sort()
-
-  let standard = Object.assign(result.standard, {globals})
-
-  result = Object.assign(result, {
-    scripts: sortObj(scripts),
-    standard: sortObj(standard)
+  return Object.assign(info, {
+    scripts: sortObj(scripts)
   })
-
-  return result
 }
 
-function typeGeneric (info, props) {
-  let result = Object.assign({}, info)
+function typeGeneric (info, {command}) {
+  /* Define cli command only if it's provided */
+  let bin = command ? {
+    [command]: 'dist/cli'
+  } : {}
 
-  if (props.command) {
-    result = Object.assign(result, {
-      bin: {
-        [props.command]: 'dist/cli'
-      }
-    })
-  }
-
-  let scripts = Object.assign(result.scripts, {
+  let scripts = Object.assign(info.scripts, {
     'start': 'babel-node src/index'
   })
 
-  result = Object.assign(result, {
+  return Object.assign(info, {
+    bin,
     scripts: sortObj(scripts)
   })
-
-  return result
 }
 
 function typeServer (info, props) {
-  let result = Object.assign({}, info)
-
-  let scripts = Object.assign(result.scripts, {
+  let scripts = Object.assign(info.scripts, {
     'serve': 'node dist/index',
     'start': 'nodemon src/index --exec babel-node'
   })
 
-  result = Object.assign(result, {
+  return Object.assign(info, {
     scripts: sortObj(scripts)
   })
-
-  return result
 }
